Handle terms in the top-level of guards properly
Expressions in guards such as:

f() when [1,2,3] ->
    ok.

would cause the debugger to crash when attempting to interpret the
module containing the expressions. Other kind of constants such as:

f() when 42 ->
    ok.

were converted to an invalid internal format ({integer,Line,42}
instead of {value,Line,42}), but that happened to work because
because anything not equal to 'true' (even a crash) was interpreted
as 'false'.

Make sure to handle all possible expressions and convert them
directly to {value,Line,false}. Remove the special handling of
the atom 'true' in and_guard/1 since it is no longer needed.