debugger: View terms as unicode-strings
Changed the displayed variables to be printed with ~tp to show default
all lists and utf-8 binaries as unicode strings.

If the user want to see the content without unicode strings he can click
on them to see the content as integers, if not it only contains latin1
because ~p converts that.

There is currently no way to display it as lists with indentation,
so I choose to keep it as it was.